requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.5
django-classy-tags==3.0.1
django-cors-headers==3.13.0
gunicorn==20.1.0
Pillow==9.1.1
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
whitenoise==6.2.0